% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg}
\alias{ridgereg}
\title{A RC class for ridgereg regression}
\description{
A RC class for ridgereg regression
}
\section{Fields}{

\describe{
\item{\code{formula}}{The formula object containing depedent and independent variables}

\item{\code{data}}{A data frame object to apply the multiple linear regression to}

\item{\code{lambda}}{numeric value which is default set as Zer0
a<-ridgereg$new(formula=Petal.Length ~ Sepal.Width + Sepal.Length,data=iris)
plot(ridgereg$new(formula=Petal.Length ~ Sepal.Width + Sepal.Length,data=iris))}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{This function returns the vector of beta coefficients}

\item{\code{initialize(formula, data, lambda = 0, normalize = FALSE)}}{This function acts as constructor}

\item{\code{predict()}}{This function returns the vector of calculated fitted values}

\item{\code{print()}}{This function prints the formula and dataset name as well as the calculated coefficients}
}}

